batesste-s3-dyndns: Add an AWS S3-based dynamic DNS system
In order to be able to access my home network from outside the house I
need to know the WAN IP. This is done via a systemd timer that calls a
bash script that obtains the WAN IP and then, if needed, updates an
AWS S3 bucket I can access from anywhere.

We also add a CI workflow to test this on GitHub.