backups: Add a systemd based backup method
In order to do backups of the servers in my homelab we add a set of
scripts that we can install as a systemd service. These scripts use
dd, pigz and mount-s3 to copy a compressed version of a given set of
block devices to an AWS S3 bucket.

We also add a GitHub CI test for this.